+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `laravel-imagekit` will be documented in this file.
 
-## Unreleased
+## v0.8.0
 
 ### Added